Add files via upload
SMART BIN USING ARDUINO UNO

This project involves developing a smart dustbin that uses an Arduino UNO, ultrasonic sensors, and a motor to enable contactless waste disposal. The smart dustbin detects when an object is nearby and automatically opens its lid, minimizing physical contact and promoting hygiene. This innovation is particularly relevant in the context of the pandemic, where reducing contact with surfaces helps prevent the spread of viruses. Overall, the Smart Bin project aims to enhance cleanliness and hygiene in both public and private spaces.

Environmental Concerns: The project addresses the growing environmental issues due to rapid population increase and the consequent rise in garbage production.
Smart Dustbin: Development of a smart dustbin that enables contactless waste disposal using technology.
Components Used: Utilizes an Arduino UNO, ultrasonic sensors, and a motor for its operation.
Functionality: Detects the presence of objects nearby and automatically opens the lid of the dustbin, reducing the need for physical contact.
Hygiene Promotion: Aims to promote hygiene by minimizing physical contact, which is especially crucial during the ongoing pandemic to prevent the spread of viruses.
Application: Designed for use in both public and private spaces to improve cleanliness and hygiene.